+++ tests/testthat/test-ablation-cmdline.R
@@ -7,8 +7,8 @@ withr::with_output_sink("test-ablation-cmdline.Rout", {
   
   test_that("--log-file=sann.rda", {
     logfile <- system.file(package="irace", mustWork = TRUE, file.path("exdata","sann.rda"))
-    outfile <- tempfile(pattern = "log-ablation", fileext = ".Rdata")
-    plotfile <- tempfile(pattern = "ablation", fileext = ".pdf")
+    outfile <- withr::local_tempfile(pattern = "log-ablation", fileext = ".Rdata")
+    plotfile <- withr::local_tempfile(pattern = "ablation", fileext = ".pdf")
     expect_no_warning(
       ablation_cmdline(paste0("--log-file=", logfile, " -o ", outfile, " -p ", plotfile)))
   })